Find Missing Positive Integer Method 2 (ACTUAL)
Space complexity is constant as all changes are made in the original vector.
Logic: Consider a vector only of positive integers, A. The first missing number has to belong to 1..A.size(). Therefore while traversing A, if A[i]-1<A.size(), we mark A[element-1] as negative (element=A[i]). Hence if a positive number is missing, the element at it's given index will remain positive and can be found easily.
SInce we have non-positive integers too, we rearrange A with negatives to one side.