Min XOR Value
It is observed that for any 3 given integers A<B<C; XOR(A,C) will always be greater than XOR(A,B) or XOR(A,C). Therefore, a simple approach we can use for the given set of integers is to sort the integers and calculate consequent pairwise XOR. The minimum will be found among these.
P.S.: Doesn't seem to give good results when simultaneously comparing negative and positive integers, but then again, handling negative integers on binary can be tricky.