Minimize the Absolute Difference
We start with the last elements of each of the vectors. Since sorted, these are the largest elements of the arrays. Thereafter, if we want to decrease the difference, we will try reducing the max of the three, i.e., going to its previous elements and if the new difference is less, update the absolute difference value.