Evaluate Expression
The idea is that a operator works on it's previous two operands. So we store the elements in a stack and when we encounter a operator, we apply in on the last two elements of the stack, while popping them. Then we store the result in the stack. Finally we return the top element of the stack.