Kth Smallest Element In Tree
In the first one I did full in-order traversal of tree, storing the result in an array and returning the kth element. Here however, I have maintained a counter to stop and return as soon as we reach kth element. Saved both time and space as we no longed require an array to store the result.