Flatten Binary Tree to Linked List (Inefficient)
My approach was to perform a in-order traversal while creating a new tree (,i.e., Linked List). However this caused Memory Limit to be exceeded. 
Alternate approach is a pre-order traversal whilst altering the existing tree. Since the right nodes are already stored, the rest of the tree remains intact.